refactor(rag): move store-and-link helpers into app/rag/ingestion.py
